+++ linkture.py
@@ -26,41 +26,47 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWAregex.
 """
 
-VERSION = '1.5.0'
+VERSION = '2.0.0'
 
-
-import argparse, json, regex
+import argparse, json, regex, sqlite3
 import pandas as pd
 
 from pathlib import Path
 from unidecode import unidecode
 
+available_languages = ('Chinese', 'Danish', 'Dutch', 'English', 'French', 'German', 'Greek', 'Italian', 'Japanese', 'Korean', 'Norwegian', 'Polish', 'Portuguese', 'Russian', 'Spanish')
 
 class Scriptures():
 
-    def __init__(self, language='English', translate=None, form=0, rewrite=False):
-        if language not in ('Chinese', 'Danish', 'Dutch', 'English', 'French', 'German', 'Greek', 'Italian', 'Japanese', 'Korean', 'Norwegian', 'Polish', 'Portuguese', 'Russian', 'Spanish'):
-            return
+    def __init__(self, language='English', translate=None, form=None):
+        if language not in available_languages:
+            raise ValueError("Indicated source language is not an option!")
         if translate:
-            if translate not in ('Chinese', 'Danish', 'Dutch', 'English', 'French', 'German', 'Greek', 'Italian', 'Japanese', 'Korean', 'Norwegian', 'Polish', 'Portuguese', 'Russian', 'Spanish'):
-                return
+            if translate not in available_languages:
+                raise ValueError("Indicated translation language is not an option!")
         else:
             translate = language
-        self.bn = {}
-        self.rewrite = rewrite
+        self.rewrite = bool((language != translate) or form)
+        if form == "full":
+            form = 3
+        elif form == "standard":
+            form = 4
+        elif form == "official":
+            form = 5
+        else:
+            form = 3
+        self.src_book_names = {}
         path = Path(__file__).resolve().parent
 
-        self.books = ['Bible']
-        with open(path / 'res/books.json', 'r', encoding='UTF-8') as json_file:
-            b = json.load(json_file)
-        for row in b[translate]:
-            names = row[1].split(', ')
-            self.books.insert(row[0], names[form])
-        for row in b[language]:
-            names = row[1].split(', ')
-            for item in names:
-                normalized = unidecode(item.replace(' ', '').replace('.', '').replace('-', '').upper())
-                self.bn[normalized] = row[0]
+        self.tr_book_names = ['Bible']
+        con = sqlite3.connect(path / 'res/resources.db')
+        cur = con.cursor()
+        for rec in cur.execute(f"SELECT * FROM Books WHERE Language = '{translate}';").fetchall():
+            self.tr_book_names.insert(rec[2], rec[form])
+        for rec in cur.execute(f"SELECT * FROM Books WHERE Language = '{language}';").fetchall():
+            for i in range(3,6):
+                normalized = unidecode(rec[i].replace(' ', '').replace('.', '').replace('-', '').upper())
+                self.src_book_names[normalized] = rec[2]
         with open(path / 'res/custom.json', 'r', encoding='UTF-8') as json_file:
             b = json.load(json_file)
         if language in b.keys():
@@ -68,48 +74,69 @@ class Scriptures():
                 names = row[1].split(', ')
                 for item in names:
                     normalized = unidecode(item.replace(' ', '').replace('.', '').replace('-', '').upper())
-                    self.bn[normalized] = row[0]
-
-        self.br = pd.read_csv(path / 'res/ranges.csv', delimiter='\t')
+                    self.src_book_names[normalized] = row[0]
+        self.ranges = pd.read_sql_query("SELECT * FROM Ranges;", con)
+        cur.close()
+        con.close()
 
         self.bk_ref = regex.compile(r'(\d?(?:\s?[\p{L}\.-]+)+)\s?(.*)') # CHECK: not tested with non-Latin characters
-        self.ch_v_ch_v = regex.compile(r'(\d+)\s*:\s*(\d+)\s*[-\u2013\u2014]\s*(\d+)\s*:\s*(\d+)')
+        # self.bk_ref = regex.compile(r'((?:\d{0,1}|[Ii]{0,3})[\.\-\s]?\p{Lu}[\p{L}.\-]+)(\d.*)') # CHECK: not tested with non-Latin characters
+        # self.bk_ref = regex.compile(r'((?:\d{0,1}|[Ii]{0,3})[\.\-\s]?\p{Lu}[\p{L}.\-]+)')
+        self.ch_v_ch_v = regex.compile(r'(\d+)\s*:\s*(\d+)[-\u2013\u2014\s]+(\d+)\s*:\s*(\d+)')
         self.ch_v_v = regex.compile(r'(\d+)\s*:\s*(\d+)\s*[-\u2013\u2014]\s*(\d+)')
         self.ch_v = regex.compile(r'(\d+)\s*:\s*(\d+)')
         self.ch_ch = regex.compile(r'(\d+)\s*[-\u2013\u2014]\s*(\d+)')
         self.ch_ = regex.compile(r'(\d+)')
-        self.v_v = regex.compile(r'(?=(\d+\s*),(\s*\d+))')
+        self.v_v = regex.compile(r'(?=(\d+\s*),(\s*\d+))') # NOTE: what if it's like 2:3,4,5? Should be reduced to 2:3-5
+        self.d_d_d = regex.compile(r'(\d+)[\s\,]+(\d+)[\s\,]+(\d+)')
+        self.d_d = regex.compile(r'(\d+)[-\u2013\u2014\s]+(\d+)')
         self.vv = regex.compile(r'(?<!:)(\d+)\s*-\s*(\d+)')
         self.dd = regex.compile(r'(\d+)\s*-\s*(\d+)\s*(?!:)')
+        self.scrpt = regex.compile(r'((?:\d{0,1}|[Ii]{0,3})[\.\-\s]?\p{Lu}[\p{L}.\-]+[:​\.\-\u2013\u2014\d,\s;]*(?<!;\s)\d)')
 
-    def _check_book(self, book):
-        bk = unidecode(book).upper().replace(' ', '').replace('.', '').replace('-', '')
-        if bk not in self.bn:
+    def _check_book(self, bk_name):
+        bk_name = unidecode(bk_name).upper().replace(' ', '').replace('.', '').replace('-', '')
+        if bk_name not in self.src_book_names:
             return None, 0
         else:
-            book = self.bn[bk]
-        return self.br.loc[(self.br.Book == book) & (self.br.Chapter.isnull()), ['Book', 'Last']].values[0]
+            bk_num = self.src_book_names[bk_name]
+        return self.ranges.loc[(self.ranges.Book == bk_num) & (self.ranges.Chapter.isnull()), ['Book', 'Last']].values[0]
 
-    def _process_scripture(self, scripture):
-        result = self.bk_ref.search(scripture)
-        if result:
-            bk, rest = result.group(1), result.group(2).lstrip()
-            bn, last = self._check_book(bk)
-            if not bn:
-                return '', '', None, -1
-            if rest == "":
-                vss = self.br.loc[(self.br.Book == bn) & (self.br.Chapter == last), ['Last']].values[0][0]
-                rest = f"1:1-{last}:{vss}"
-            for result in self.v_v.findall(rest):
-                if int(result[1]) - int(result[0]) == 1:
-                    rest = rest.replace(f"{result[0]},{result[1]}", f"{result[0].rstrip()}-{result[1].lstrip()}")
-            if bn:
-                if self.rewrite:
-                    bk = self.books[bn]
-                return f"{bk} ", rest, bn, last
-        return '', '', None, 0
+    # def _process_scripture(self, scripture): # TODO: returns rewritten name, not original
 
-    def link_scripture(self, scripture):
+    #     def check_book(bk_name):
+    #         bk_name = unidecode(bk_name).upper().replace(' ', '').replace('.', '').replace('-', '')
+    #         if bk_name not in self.src_book_names:
+    #             return None, 0
+    #         else:
+    #             bk_num = self.src_book_names[bk_name]
+    #         return self.ranges.loc[(self.ranges.Book == bk_num) & (self.ranges.Chapter.isnull()), ['Book', 'Last']].values[0]
+
+    #     result = self.bk_ref.search(scripture)
+    #     if result:
+    #         bk_name, rest = result.group(1), result.group(2).lstrip()
+    #         bk_num, last = check_book(bk_name)
+    #         if not bk_num:
+    #             return '', '', None, -1
+    #         if rest == "":
+    #             vss = self.ranges.loc[(self.ranges.Book == bk_num) & (self.ranges.Chapter == last), ['Last']].values[0][0]
+    #             rest = f"1:1-{last}:{vss}"
+    #         for result in self.v_v.findall(rest):
+    #             if int(result[1]) - int(result[0]) == 1:
+    #                 rest = rest.replace(f"{result[0]},{result[1]}", f"{result[0].rstrip()}-{result[1].lstrip()}")
+    #         if bk_num:
+    #             return bk_name, rest, bk_num, last
+    #     return '', '', None, 0
+
+    def list_scriptures(self, text):
+        return regex.findall(self.scrpt, text)
+
+    def tag_scriptures(self, text):
+        def r(match):
+            return "{{" + match.group(1) + "}}"
+        return regex.sub(self.scrpt, r, text)
+
+    def link_scripture(self, scripture, prefix='<a href="http://', suffix='" >'):
 
         def process_verses(chunk, book, multi):
             b = str(book)
@@ -142,7 +169,7 @@ class Scriptures():
                     ch1 = result.group(1)
                     v1 = '1'
                     ch2 = result.group(2)
-                    v2 = str(self.br.loc[(self.br.Book == book) & (self.br.Chapter == int(ch2)), ['Last']].values[0][0])
+                    v2 = str(self.ranges.loc[(self.ranges.Book == book) & (self.ranges.Chapter == int(ch2)), ['Last']].values[0][0])
                 else:
                     ch1 = '1'
                     v1 = result.group(1)
@@ -156,7 +183,7 @@ class Scriptures():
                     ch1 = result.group(1)
                     v1 = '1'
                     ch2 = ch1
-                    v2 = str(self.br.loc[(self.br.Book == book) & (self.br.Chapter == int(ch2)), ['Last']].values[0][0])
+                    v2 = str(self.ranges.loc[(self.ranges.Book == book) & (self.ranges.Chapter == int(ch2)), ['Last']].values[0][0])
                     return f"{b}:{ch1}:{v1}-{b}:{ch2}:{v2}", 0
                 else:
                     ch1 = '1'
@@ -175,34 +202,132 @@ class Scriptures():
                     txt = txt.replace(f"{result[0]}-{result[1]}", f"{result[0]}, {result[1]}")
             return txt
 
-        url = ''
+        output = ''
         book = ''
         for chunk in scripture.split(';'):
             try:
-                bk, rest, bn, last = self._process_scripture(chunk)
-                if last == -1:
-                    url = url + '; ' + '{{' + chunk.strip() + '}}'
-                    continue
-                if not bn:
-                    bk, rest, bn, last = self._process_scripture(book + chunk)
-                if bk.strip() != book.strip():
-                    book = bk
+                bk_name, rest, bk_num, last = self._process_scripture(chunk)
+                # if last == -1:
+                #     output = output + '; ' + '{{' + chunk.strip() + '}}'
+                #     continue
+                if not bk_num:
+                    bk_name, rest, bk_num, last = self._process_scripture(book + chunk)
+                if bk_name.strip() != book.strip():
+                    book = bk_name
                 else:
-                    bk = ''
-                chap = 0
+                    bk_name = ''
+                ch = 0
                 for bit in rest.split(','):
-                    if chap:
-                        link, chap = process_verses(f"{chap}:{bit}", bn, last-1)
-                        url += ', '
+                    if ch:
+                        link, ch = process_verses(f"{ch}:{bit}", bk_num, last-1)
+                        output += ', '
                     else:
-                        link, chap = process_verses(bit, bn, last-1)
-                        url += '; '
-                    processed_chunk = f"{bk}{undo_series(bit).lstrip()}"
-                    url += f'<a href="jwpub://b/NWTR/{link}" class="b">{processed_chunk.strip()}</a>'
-                    bk = ''
+                        link, ch = process_verses(bit, bk_num, last-1)
+                        output += '; '
+                    processed_chunk = f"{bk_name}{undo_series(bit).lstrip()}"
+                    output += f'{prefix}{link}{suffix}{processed_chunk.strip()}</a>'
+                    bk_name = ''
             except:
-                url += "{{" + chunk + "}}"
-        return url.strip(' ;,')
+                output += "{{" + chunk + "}}"
+        return output.strip(' ;,')
+
+
+
+    # def _process_scripture(self, scripture):
+
+    #     def check_book(bk_name):
+    #         bk_name = unidecode(bk_name).upper().replace(' ', '').replace('.', '').replace('-', '')
+    #         if bk_name not in self.src_book_names:
+    #             return None, 0
+    #         else:
+    #             bk_num = self.src_book_names[bk_name]
+    #         return self.ranges.loc[(self.ranges.Book == bk_num) & (self.ranges.Chapter.isnull()), ['Book', 'Last']].values[0]
+
+    #     result = self.bk_ref.search(scripture)
+    #     if result:
+    #         bk_name = result.group(1).strip()
+    #         rest = scripture[len(bk_name):]
+    #         bk_num, last = check_book(bk_name)
+    #     else:
+    #         bk_name = None
+    #         rest = scripture
+    #         bk_num = None
+    #         last = None
+    #     while True:
+    #         result = self.d_d_d.search(rest)
+    #         if not result:
+    #             break
+    #         end = result.group(3)
+    #         start = result.group(1)
+    #         if int(end) - int(start) == 2:
+    #             rest = regex.sub(start+r'[\s\d\,]+'+end, f"{start.rstrip()}-{end.lstrip()}", rest)
+    #     if not bk_num:
+    #         return '', '', None, -1
+
+    #     if bk_num:
+    #         return bk_name, rest, bk_num, last
+    #     return '', '', None, 0
+
+
+    def rewrite_scripture(self, scripture):
+
+        def reform_series(txt): # rewrite comma-separated consecutive sequences as (1, 2, 3) as ranges (1-3) and consecutive ranges (1-2) as comma-separated sequences (1, 2)
+            vv_v = regex.compile(r'(\d+)[\s\,]+(\d+)[-\u2013\u2014\s]+(\d+)')
+            v_vv = regex.compile(r'(\d+)[-\u2013\u2014\s]+(\d+)[\s\,]+(\d+)')
+            found = True
+            while found:
+                found = False
+                for result in self.d_d_d.finditer(txt):
+                    end = result.group(3)
+                    start = result.group(1)
+                    if int(end) - int(start) == 2:
+                        found = True
+                        txt = regex.sub(result.group(), f"{start}-{end}", txt)
+                for result in v_vv.finditer(txt):
+                    end = result.group(3)
+                    mid = result.group(2)
+                    start = result.group(1)
+                    if int(end) - int(mid) == 1:
+                        found = True
+                        txt = regex.sub(result.group(), f"{start}-{end}", txt)
+                for result in vv_v.finditer(txt):
+                    end = result.group(3)
+                    mid = result.group(2)
+                    start = result.group(1)
+                    if int(mid) - int(start) == 1:
+                        found = True
+                        txt = regex.sub(result.group(), f"{start}-{end}", txt)
+                for result in self.d_d.finditer(txt):
+                    end = result.group(2)
+                    start = result.group(1)
+                    if int(end) - int(start) == 1:
+                        found = True
+                        txt = regex.sub(result.group(), f"{start},{end}", txt)
+                for result in self.ch_v_ch_v.finditer(txt):
+                    sc = result.group(1)
+                    sv = result.group(2)
+                    ec = result.group(3)
+                    ev = result.group(4)
+                    if int(sc) == int(ec):
+                        found = True
+                        txt = regex.sub(result.group(), f"{sc}:{sv}-{ev}", txt)
+            return txt
+
+        result = self.bk_ref.search(scripture) # get book name
+        if result:
+            bk_name, rest = result.group(1).strip(), result.group(2).strip()
+            bk_num, last = self._check_book(bk_name)
+        else:
+            return scripture
+        if self.rewrite and bk_name and bk_num:
+            output = self.tr_book_names[bk_num]+' '
+        else:
+            output = bk_name+' '
+        for chunk in rest.replace(' ', '').split(';'): # process the rest
+            chunk = reform_series(chunk)
+            output += chunk.lstrip()+'; '
+        return output.replace(',', ', ').strip(' ;,')
+
 
     def code_scripture(self, scripture):
 
@@ -237,7 +362,7 @@ class Scriptures():
                     ch1 = result.group(1).zfill(3)
                     v1 = '001'
                     ch2 = result.group(2).zfill(3)
-                    v2 = str(self.br.loc[(self.br.Book == book) & (self.br.Chapter == int(ch2)), ['Last']].values[0][0]).zfill(3)
+                    v2 = str(self.ranges.loc[(self.ranges.Book == book) & (self.ranges.Chapter == int(ch2)), ['Last']].values[0][0]).zfill(3)
                 else:
                     ch1 = '001'
                     v1 = result.group(1).zfill(3)
@@ -251,7 +376,7 @@ class Scriptures():
                     ch1 = result.group(1).zfill(3)
                     v1 = '001'
                     ch2 = ch1
-                    v2 = str(self.br.loc[(self.br.Book == book) & (self.br.Chapter == int(ch2)), ['Last']].values[0][0]).zfill(3)
+                    v2 = str(self.ranges.loc[(self.ranges.Book == book) & (self.ranges.Chapter == int(ch2)), ['Last']].values[0][0]).zfill(3)
                     return (b+ch1+v1, b+ch2+v2), 0
                 else:
                     ch1 = '001'
@@ -263,26 +388,33 @@ class Scriptures():
         series = []
         book = ''
         for chunk in scripture.split(';'):
-            try:
-                bk, rest, bn, last = self._process_scripture(chunk)
-                if last == -1:
-                    continue
-                if not bn:
-                    bk, rest, bn, last = self._process_scripture(book + chunk)
-                    bk = ''
+            try: # TODO: needs fixing:
+                bk_name, rest, bk_num, last = self._process_scripture(chunk)
+                # if last == -1:
+                #     continue
+                if not bk_num:
+                    bk_name, rest, bk_num, last = self._process_scripture(book + chunk)
+                    bk_name = ''
                 else:
-                    book = bk
-                chap = 0
+                    book = bk_name.strip()
+                ch = 0
                 for bit in rest.split(','):
-                    if chap:
-                        link, chap = code_verses(f"{chap}:{bit}", bn, last-1)
+                    if ch:
+                        link, ch = code_verses(f"{ch}:{bit}", bk_num, last-1)
                     else:
-                        link, chap = code_verses(bit, bn, last-1)
+                        link, ch = code_verses(bit, bk_num, last-1)
                     series.append(link)
-                    bk = ''
+                    bk_name = ''
             except:
                 pass
         return series
+
+    def code_scriptures(self, text):
+        lst = []
+        for i in self.list_scriptures(text):
+            for c in self.code_scripture(i):
+                lst.append(c)
+        return lst
 
     def decode_scripture(self, reference=[]):
         scriptures = ''
@@ -298,57 +430,50 @@ class Scriptures():
             ev = int(end[5:])
             if not ((0 < sb <= 66) & (sb == eb)): # book out of range
                 continue
-            if not (0 < sc <= ec <= self.br.loc[(self.br.Book == sb) & (self.br.Chapter.isnull()), ['Last']].values[0]): # chapter(s) out of range
+            if not (0 < sc <= ec <= self.ranges.loc[(self.ranges.Book == sb) & (self.ranges.Chapter.isnull()), ['Last']].values[0]): # chapter(s) out of range
                 continue
-            if not ((0 < sv <= self.br.loc[(self.br.Book == sb) & (self.br.Chapter == sc), ['Last']].values[0]) & (0 < ev <= self.br.loc[(self.br.Book == sb) & (self.br.Chapter == ec), ['Last']].values[0])): # verse(s) out of range
+            if not ((0 < sv <= self.ranges.loc[(self.ranges.Book == sb) & (self.ranges.Chapter == sc), ['Last']].values[0]) & (0 < ev <= self.ranges.loc[(self.ranges.Book == sb) & (self.ranges.Chapter == ec), ['Last']].values[0])): # verse(s) out of range
                 continue
-            bk = self.books[sb]
-            if self.br.loc[(self.br.Book == sb) & (self.br.Chapter.isnull()), ['Last']].values[0] == 1:
+            bk_name = self.tr_book_names[sb]
+            if self.ranges.loc[(self.ranges.Book == sb) & (self.ranges.Chapter.isnull()), ['Last']].values[0] == 1:
                 ch = ' '
             else:
                 ch = f" {sc}:"
             if start == end:
-                scripture = f"{bk}{ch}{sv}"
+                scripture = f"{bk_name}{ch}{sv}"
             else:
                 if sc == ec:
                     if ev - sv == 1:
-                        scripture = f"{bk}{ch}{sv}, {ev}"
+                        scripture = f"{bk_name}{ch}{sv}, {ev}"
                     else:
-                        scripture = f"{bk}{ch}{sv}-{ev}"
+                        scripture = f"{bk_name}{ch}{sv}-{ev}"
                 else:
-                    scripture = f"{bk}{ch}{sv}-{ec}:{ev}"
+                    scripture = f"{bk_name}{ch}{sv}-{ec}:{ev}"
             scriptures += f"; {scripture}"
         return scriptures.lstrip(" ;")
-
-    def rewrite_scripture(self, scripture):
-        return self.decode_scripture(self.code_scripture(scripture))
 
 
 def _main(args):
 
-    def replacement(match):
+    def r(match):
         group = match.group(1).strip('{}')
         if not args['quiet']:
             print(f'...Processing "{group}"')
         if args['link']:
-            return s.link_scripture(group)
+            return s.link_scripture(group, '<a href="jwpub://b/NWTR/', '" class="b">')
         elif args['range']:
             return str(s.code_scripture(group))
         else:
             return s.rewrite_scripture(group)
 
-    rewrite = False
-    form = 0
+    form = None
     if args['standard']:
-        form = 1
-        rewrite = True
+        form = 'standard'
     elif args['official']:
-        form = 2
-        rewrite = True
+        form = 'official'
     elif args['full']:
-        form = 0
-        rewrite = True
-    s = Scriptures(args['language'], args['translate'], form, rewrite)
+        form = 'full'
+    s = Scriptures(args['language'], args['translate'], form)
     m = regex.compile(r'({{.*?}})')
 
     if args['f']:
@@ -358,15 +483,16 @@ def _main(args):
         with open(args['f'][0], 'r', encoding='UTF-8') as f:
             txt = f.read()
     else:
-        txt = "{{" + args['s'] + "}}"
+        txt = args['s']
 
-    txt2 = regex.sub(m, replacement, txt)
+    txt = s.tag_scriptures(txt)
+    txt = regex.sub(m, r, txt)
 
     if args['f']:
         with open(args['f'][1], 'w', encoding='UTF-8') as f:
-            f.write(txt2)
+            f.write(txt)
     else:
-        print(txt2)
+        print(txt)
 
 if __name__ == "__main__":
     PROJECT_PATH = Path(__file__).resolve().parent
@@ -380,16 +506,15 @@ if __name__ == "__main__":
     mode.add_argument('-f', metavar=('in-file', 'out-file'), nargs=2, help='work with files (UTF-8)')
     mode.add_argument('-s', metavar='reference', help='process "reference; reference; etc."')
 
-    parser.add_argument('--language', default='English', choices=['Chinese', 'Danish', 'Dutch', 'English', 'French', 'German', 'Greek', 'Italian', 'Japanese', 'Korean', 'Norwegian', 'Polish', 'Portuguese', 'Russian', 'Spanish'], help='indicate source language for book names (English if unspecified)')
-    parser.add_argument('--translate', default='English', choices=['Chinese', 'Danish', 'Dutch', 'English', 'French', 'German', 'Greek', 'Italian', 'Japanese', 'Korean', 'Norwegian', 'Polish', 'Portuguese', 'Russian', 'Spanish'], help='indicate output language for book names (same as source if unspecified)')
-
+    parser.add_argument('--language', default='English', choices=available_languages, help='indicate source language for book names (English if unspecified)')
+    parser.add_argument('--translate', choices=available_languages, help='indicate output language for book names (same as source if unspecified)')
     format_group = parser.add_argument_group('output format (optional)', 'if provided, book names will be rewritten accordingly:')
-    form = format_group.add_mutually_exclusive_group(required=False)
-    form.add_argument('--full', action='store_true', help='output as full name - default (eg., "Genesis")')
-    form.add_argument('--official', action='store_true', help='output as official abbreviation (eg., "Ge")')
-    form.add_argument('--standard', action='store_true', help='output as standard abbreviation (eg., "Gen.")')
+    formats = format_group.add_mutually_exclusive_group(required=False)
+    formats.add_argument('--full', action='store_true', help='output as full name - default (eg., "Genesis")')
+    formats.add_argument('--official', action='store_true', help='output as official abbreviation (eg., "Ge")')
+    formats.add_argument('--standard', action='store_true', help='output as standard abbreviation (eg., "Gen.")')
 
-    type_group = parser.add_argument_group('type of conversion', 'if not specified, references are simply rewritten according to chosen output format:')
+    type_group = parser.add_argument_group('type of conversion', 'if not specified, references are simply rewritten according to chosen (or default) output format:')
     tpe = type_group.add_mutually_exclusive_group(required=False)
     tpe.add_argument('-l', '--link', action='store_true', help='create jwpub link(s)')
     tpe.add_argument('-r', '--range', action='store_true', help='create range list')